requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
openai>=1.0.0
anthropic>=0.18.0
lxml>=4.9.0
httpx>=0.25.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
aiohttp>=3.9.0
tenacity>=8.2.0
